requirements.txt
flask
flask-login
flask-sqlalchemy
flask-wtf
werkzeug
python-dotenv